backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
openai
razorpay
reportlab
weasyprint
jinja2
